refactor: Move all source files into 'src/'
This patch moves all source files into a common directory to simplify
understanding the project structure, and the scopes of build tools.

BREAKING CHANGE: Modifies the location of 'config.ts'.
