+++ src/controllers/legend-controller.ts
