+++ src/job.ts
@@ -2,7 +2,7 @@ import { CanteenPlan, DateSpec, fetchMensa } from 'ka-mensa-fetch'
 import { group } from 'group-items'
 import moment from 'moment'
 import ms from 'ms'
-import config from '../config.js'
+import config from './config.js'
 import { getSessionCookie } from './get-session-cookie.js'
 import { Cache } from './cache.js'
 import { logger } from './logger.js'
