+++ src/routes/default.ts
@@ -1,6 +1,6 @@
 import { Router } from 'express'
-import { Cache } from '../lib/cache.js'
-import { createHandler } from '../lib/create-handler.js'
+import { Cache } from '../cache.js'
+import { createHandler } from '../create-handler.js'
 
 /**
  * Create the router for retrieving API status information.
