+++ test/cache.test.ts
@@ -1,4 +1,4 @@
-import { Cache } from '../lib/cache.js'
+import { Cache } from '../src/cache.js'
 import { MemoryAdapter } from 'fs-adapters'
 import chai, { expect } from 'chai'
 import chaiAsPromised from 'chai-as-promised'
