+++ tsconfig.json
@@ -7,9 +7,7 @@
     "strict": true,
     "outDir": "./dist"
   },
-  "exclude": [
-    "test",
-    "dist",
-    "node_modules"
+  "include": [
+    "src"
   ]
 }
